requirements.txt
# need to install module
wxpy
pillow